+++ keyboards/for_order_status.py
@@ -1,0 +1,9 @@
+from aiogram.types import InlineKeyboardMarkup, ReplyKeyboardMarkup
+from aiogram.utils.keyboard import InlineKeyboardBuilder, ReplyKeyboardBuilder
+
+
+def get_authorize_kb() -> ReplyKeyboardMarkup:
+    kb = ReplyKeyboardBuilder()
+    kb.button(text='Авторизоваться', request_contact=True)
+
+    return kb.as_markup()
